+++ Changelog.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Add support for specifying multiple subject expressions with the RSpec integration. [#1125](https://github.com/mbj/mutant/pull/1125)
+
 # v0.10.8 2020-11-22
 
 * Add support for process abort as coverage condition.